Avoid duplicates in command intellisense
The dictionnary of commands completion is indexed by the signature of
the command. We do not add new entries with a key matching the command
name of an already existing entry. This is usefull to avoid having one
entry without arguments with we already entries with arguments. This is
an attempt to sanitize the completion list a bit.