Make citation intellisense local to a root file
When parsing bib files, we keep track of the root file to make sure the
citations appear in the intellisense menu only for the LaTeX project
with the correct root file. Because bib files may be large, we do not
reset citation data at every change of root file unlike what is done
with tex files.